refactor: CoreUI 모듈 네이밍 재정의
시뮬레이션 빌드 시 "Thread 1: "-[CUICatalog initWithName:fromBundle:error:]: unrecognized selector sent to instance 0x600002609320" 에러가 발생함. 해당 에러는 UIKit이 Asset Catalog(.xcassets)를 로딩하려다 번들/형식이 맞지 않아 내부 메서드를 잘못 호출하면서 발생하는 크래시로서, 기존에 사용하던 Core 모듈 속 CoreUI가 이미 Swift에서 사용중이기 때문에 발생하는 에러로 판단(일종의 '예약어')

CoreUI를 CoreCommonUI으로 수정하여 해결하였으며, 테스트 관련한 코드들도 재위치에 맞춰서 폴더링 정리함. 이 과정에서 DesignSystem의 의존성을 제거하고 CoreCommonUI에 Design System 을 가지도록 의존성 방향 수정
