+++ Projects/Core/CoreCommonUI/Sources/DSButton.swift
@@ -1,36 +1,16 @@
 //
-//  DesignSystemTemp.swift
-//  Manifests
+//  DSButton.swift
+//  CoreCommonUI
 //
-//  Created by 박서연 on 11/27/25.
+//  Created by 박서연 on 12/18/25.
+//  Copyright © 2025 linda. All rights reserved.
 //
 
 import UIKit
 
-public enum DesignSystemColor {
-    public static let primary = UIColor(red: 0.10, green: 0.45, blue: 0.95, alpha: 1.0)
-    public static let background = UIColor(white: 0.97, alpha: 1.0)
-    public static let danger = UIColor(red: 0.95, green: 0.15, blue: 0.15, alpha: 1.0)
-}
-
-public enum DesignSystemTypography {
-
-    public static func title(_ size: CGFloat = 24) -> UIFont {
-        UIFont.systemFont(ofSize: size, weight: .bold)
-    }
-
-    public static func body(_ size: CGFloat = 16) -> UIFont {
-        UIFont.systemFont(ofSize: size, weight: .regular)
-    }
-
-    public static func caption(_ size: CGFloat = 12) -> UIFont {
-        UIFont.systemFont(ofSize: size, weight: .medium)
-    }
-}
-import UIKit
+import DesignSystem
 
 public final class DSButton: UIButton {
-
     public enum Style {
         case primary
         case danger
