+++ Projects/Core/CoreModels/Sources/CoreModelTest.swift
@@ -1,0 +1,9 @@
+//
+//  CoreModelTest.swift
+//  CoreModels
+//
+//  Created by 박서연 on 12/18/25.
+//  Copyright © 2025 linda. All rights reserved.
+//
+
+// CoreModelTest
