--- Projects/Core/CoreModels/Sources/CoreModelsTest.swift
@@ -1,16 +1,0 @@
-//
-//  CoreModelsTest.swift
-//  CoreManifests
-//
-//  Created by 박서연 on 11/26/25.
-//
-
-import SwiftUI
-
-public struct CoreModelsTest: View {
-    public init() { }
-    
-    public var body: some View {
-        Text("hi!")
-    }
-}
