--- Projects/Core/CoreNetwork/Sources/CoreNetworkTest.swift
@@ -1,8 +1,0 @@
-//
-//  CoreNetworkTest.swift
-//  CoreManifests
-//
-//  Created by 박서연 on 11/26/25.
-//
-
-import Foundation
