+++ Projects/DesignSystem/Sources/Color/Color.swift
@@ -1,0 +1,15 @@
+//
+//  Color.swift
+//  DesignSystem
+//
+//  Created by 박서연 on 12/18/25.
+//  Copyright © 2025 linda. All rights reserved.
+//
+
+import UIKit
+
+public enum DesignSystemColor {
+    public static let primary = UIColor(resource: .primary)
+    public static let background = UIColor(resource: .background)
+    public static let danger = UIColor(resource: .danger)
+}
