+++ Projects/Features/HomeFeature/Project.swift
@@ -14,7 +14,7 @@ let homeFeatureProject = Project.makeFeatureModule(
         .project(target: "CoreNetwork", path: "../../Core/CoreNetwork"),
         .project(target: "CoreModels", path: "../../Core/CoreModels"),
         .project(target: "CoreUtils", path: "../../Core/CoreUtils"),
-        .project(target: "CoreUI", path: "../../Core/CoreUI"),
+        .project(target: "CoreCommonUI", path: "../../Core/CoreCommonUI"),
         .project(target: "DesignSystem", path: "../../DesignSystem"),
         .external(name: "RxSwift")
     ]
